Create ATR-Relative.mq5
A version of the ATR (Average True Range) indicator based on a relative calculation instead of the normal absolute calculation. Useful for comparing the relative volatility of different assets and asset classes. Note however that this type of relative-based calculation is NOT suitable if using ATR to set for example a volatility-based stop loss. You must use the standard calculation for this type of requirement.